request-queue: Allow limiting of the request queue.
Although we added the thread pool limit to prevent from too many
requests to be stuck at the same time. This would still queue requests.
Without this limit the requests queue, in cases of continued storage
issues, could grow very large keeping a lot of actions that will have to
be performed once the queue gets released.

This patch adds the --max-queued-requests to allow limiting of the
maximum queued requests.

Change-Id: I0edb736891ec9012ce71c3098206fa95f89c5b37
Signed-off-by: Saggi Mizrahi <ficoos@gmail.com>