build: Use actual libexec dir instead of hard-coded path
Modified the copy_file function in distutils to be able to substitute
libexecdir. In the future, if needed, this could be extended to a more generic
mechanism.

It's important to note that unlike the tradition of making such
replacements at build time this is done at install time since there is
no actual build step for py files.

Change-Id: I950c0ad2c762deac1bf469cb60acc8de7066d5fc
Signed-off-by: Saggi Mizrahi <ficoos@gmail.com>
Signed-off-by: Dima Kuznetsov <dkuznets@redhat.com>