Fix busy loop when pipe is closed
When pipe is closed, read() returns 0. The code was checking if errno is
EAGAIN and entering the read loop again. This check is correct only if
read failed (return value < 0) and file descriptor has been marked
nonblocking (O_NONBLOCK).

We have seen this error randomly; I assume that errno was EAGAIN because
it was not initialized for this thread, or some other call failed with
EAGAIN before the read() call.

Change-Id: Ib301153e9c1190ee61378d914cb3ebbfc3a89608
Signed-off-by: Nir Soffer <nsoffer@redhat.com>