Fix handling of partial logs
When IOProcess received partial log line from ioprocess, it was saved
until the next log, and appended to the next data received from
iprocess. However, if the next read event was complete, the saved
partial log data was not cleared, and all logs messages from this point
included the saved partial log data.

Now partial data is handled correctly and the code was simplified to
make it easier to understand and debug.

Change-Id: I55019e79ac42d039a13d3e54ec56dd9da4cffe2f
Bug-Url: https://bugzilla.redhat.com/1275062
Signed-off-by: Nir Soffer <nsoffer@redhat.com>