Fix wrong test, leaving junk symlink at /Dsadsad
testSymlinkFail was failing only when running as non-privileged user
with EACCESS. When running as root, this operation succeeded, creating a
symlink in /Dsadsad, and the test did not fail as it should.

    $ sudo nosetests test/test.py -m testSymlinkFail
    [sudo] password for nsoffer:
    .
    ----------------------------------------------------------------------
    Ran 1 test in 0.008s

    OK
    $ ls -lht /
    total 62K
    lrwxrwxrwx.   1 root root   15 Dec 12 23:02 Dsadsad -> /I/do/not/exist

Now we try to create a symlink in non-existing directory, which will
always fail regardless of the user privileges, and the test will fail if
the operation succeeds.

Change-Id: I3b2ebb608f2c82af53fe16d6f86f7348779c0022
Signed-off-by: Nir Soffer <nsoffer@redhat.com>