ioprocess: Fix logging of file descriptor path
When closing file descriptor during startup, we use to read the link
target from /proc/self/fd/name using readlink.  If readlink failed, we
copied the string "(error)" into the wrong buffer, so we logged random
junk inside a buffer from the stack.  If readlink succeeded, we did not
terminate the buffer with a null byte (readlink does not), so again we
logged random junk from the stack.

Logging random junk from the stack can break the bindings expecting utf8
string, and it can also segfault if you are lucky.

This issue was found when running the tests on ppc64 machine:

    DEBUG|(null)|Closing unrelated fd no: 10 (/dev/urandom5]\xa3\xf0)\n

This patch ensures that we always log null terminated buffer.

Change-Id: Idde46a3ae22095ef0165ab753d2a278789a244a0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>