ioprocess: Fix logger initialization
We used to create the logger queue and register the logger handler from
the logging thread. This created a race between the first log messages
and the logger thread, and the first messages were logged via the
standard glib logging, missing the log format expected by the binding.

Now we create the logger queue and register the logger in the main
thread. All the log message will go into this queue, and they will be
submitted to the bindings when the logger thread is running.

Change-Id: I5b0edd59afa67503ede1230ac7e9e0ea5c532ab5
Signed-off-by: Nir Soffer <nsoffer@redhat.com>