binding: Wait until communication thread started
When creating an IOProcess instance, it starts a communication thread
sending and receiving commands to the underlying child process.

I think that some of the tests fail on CI and Fedora builds because of
race between sending the first command to the instance, and the actual
startup of the communication thread.

This patch synchronizes with the communication thread so after creating
an instance, the communication thread is always running and the instance
is ready for sending commands.

Change-Id: I09d3bb6b221a9b75f11fd352eacd55254a413918
Signed-off-by: Nir Soffer <nsoffer@redhat.com>