tests: Fix flaky weakref test
This test fails randomly on the CI and on Fedora build machines.

It seems that the failure is caused by using assertIsNone. If the
ioprocess instance was not garbage collected before the first assert the
instance is probably referenced by the test framework, and all further
asserts will fail, since the instance will never be released until the
end of the test.

Now we extract the instance in the test and check it manually. If the
instance is not None, we delete the local reference before the next
iteration, ensuring that nothing is referencing the instance.

The old test had a useless 30 seconds timeout. Since this test depends
on another thread running and we cannot synchronize with it, we must use
some timeout. To prevent failures on overloaded test slaves, I'm using
now 5 seconds timeout.

Since this change I never seen this test fail.

Change-Id: I7bd3d57d46db1d34a3a208defed603ae29473820
Signed-off-by: Nir Soffer <nsoffer@redhat.com>