binding: Fix race between _run and close
When the ioprocess crash, or the communication thread fail, it kills the
underlying child process and start a new one. If close is called in the
same time, we may try to wait for a thread that was not started yet.

Here is an example failure:

    2016-05-08 13:11:37,129 DEBUG   [IOProcess] (ioprocess communication (11989)) (2) Start request for method 'crash' (waitTime=2803)
    2016-05-08 13:11:37,130 ERROR   [IOProcessClient] (ioprocess communication (11989)) IOProcess failure
    Traceback (most recent call last):
      File "/builddir/build/BUILD/ioprocess-0.16.0/bindings/python/ioprocess/__init__.py", line 128, in _communicate
        raise Exception("FD closed")
    Exception: FD closed
    2016-05-08 13:11:37,132 DEBUG   [IOProcessClient] (ioprocess communication (11989)) Starting IOProcess...
    2016-05-08 13:11:37,140 DEBUG   [IOProcessClient] (ioprocess communication (11989)) Starting communication thread

To fix this race, we take a lock now when starting a new child process.
If close is called while a new process is started, it will wait on the
lock until the underlying process and the communication thread are
restarted, and then the new process and communication threads will be
stopped.

Change-Id: I3ca7675f089bdf00ad375e968dae9ed6b5caaa87
Signed-off-by: Nir Soffer <nsoffer@redhat.com>