tests: Create ioprocess in the test
We use to create ioproces instance in setUp, but some tests create their
own instance, creating unwanted noise in the log and making it hard to
debug issue in the tests. Other tests manipulate the instance timeout.

Now we create an ioprocess instance for each test with the configuration
required for the test and clean up using contextlib.closing().

Change-Id: I46723cbd8b160b0f14dfb5b373257892e42fb248
Signed-off-by: Nir Soffer <nsoffer@redhat.com>