python3: Initial port
Minimal fix to make the tests pass with Python 3. The python 3 tests
will run if python 3 is installed.

On the CI, we require now python3 on the Fedora builds, so the text
would run on both python 2 and 3.

On el7 we have very old tox that does not support skipping missing
interpreters, so we have to detect python3 in configure, and run the
tests only on python 2.7.

Change-Id: I447459159572c2a9a4007e63e53d57509751687c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>