tests: Fix test that did not test anything
testMultitask defined a test function but never run it. Looks like
accidental indentaion broke the test.

Looking at the code, it seems that the intended purpose of the tests was
testing multiple requests using same command. For this purpose, using
separate timeout for each call does not seems useful for anything, so
now the test uses timeout of 1 second for all threads.

Change-Id: Iabcd8ae7d8fea79e8086ff95541beccf0bd5a9c1
Signed-off-by: Nir Soffer <nsoffer@redhat.com>