ioprocess: Fix fd leak in truncate and touch
In commit 7dec01960213 (Add missing functionality to exported functions)
we added support for touch and truncate, leaking file descriptor 0.

The cleanup code in these functions was wrong, assuming that file
descriptor 0 will never be used, so the call to touch or truncate leak
file descriptor 0 for the entire process life.

This issue cause ioprocess to keep an open file for
/rhev/data-center/mnt/server:_path/__DIRECT_IO_TEST__. This file is not
closed even in maintenance mode, because vdsm does not terminate
ioprocess when deactivating a domain.

Change-Id: I6a080a95b21e6bef59bee0525352a3728d36183e
Bug-Url: https://bugzilla.redhat.com/1371634
Signed-off-by: Nir Soffer <nsoffer@redhat.com>