binding: Add optional ioprocess name
The management program can name each ioprocess instance. The name of the
instance is logged when creating and closing an instance.

If the name is not specified, we use ioprocess-N, where N is a serial
number, similar to how threads are named.

When staring an instance, we will see this log:

    2016-05-26 19:41:51,642 DEBUG   [IOProcessClient] (MainThread) Starting
    client ioprocess-0

When starting the ioprocess for this client, we will see this log:

    2016-05-26 19:41:51,642 DEBUG   [IOProcessClient] (MainThread) Starting
    ioprocess for client ioprocess-0

This log may show multiple times, since the communication thread may start a
new ioprocess after failures.

When closing an instance, we will see this log:

    2016-05-26 19:41:51,645 DEBUG   [IOProcessClient] (MainThread) Closing
    client ioprocess-0

Change-Id: Ia80094e166c786d3d165c03b8132389723af91d5
Signed-off-by: Nir Soffer <nsoffer@redhat.com>