build: Require ioprocess of same version
python-ioprocess was requiring ioprocess-version-release. Since release
is actually release.timestamp.githash, python-ioprocess required the
version of iprocess from the same build. Since python ioprocess is built
when building on different architectures (e.g. x86_64, ppc64), the
latest build of python-ioprocess broke the previous build, depending on
a package which does not exists in the repository.

This requirement is also wrong, as python-ioprocess must be compatible
with any ioprocess using the same major version.

We depend now on ioprocess using the same version, ignoring the release.

Change-Id: I5cd94c181e4a864b1f3bff81bcb0b5b702f4c26c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>