python: Wait for child process on close
When closing IOProcess, wait for the child process. Previously the
communication thread was starting a new thread for waiting, so the child
sometimes terminates after close() returned. This caused Vdsm's tests
to fail when the test framework is interrupted by the SIGCHLD signal.
Waiting for the child process ensures that the signal is received before
close() returned.

This change will ensure that after closing IOProcess it does not access
storage; this guarantee is critical for storage code holding a lease on
storage. The caller must not release the lease if there is a running
ioprocess that may access this storage.

Note that stuck ipprocess will cause the caller program to wait for it.
Before this change, the program never waited for it. Programs that do
not want to wait should use close(sync=False) to avoid waiting for the
child process.

Change-Id: I961da1c7dcbda6e464a23259675e7a45fdff29d8
Signed-off-by: Nir Soffer <nsoffer@redhat.com>