Replace cpopen with subprocess32
subprocess32 is a drop-in replacement of subprocess module in python 2,
improving reliability when using threads. It replaces cpopen which is a
similar solution developed for ovirt.

We require now python-subprocess32 since as it is available on all
supported platforms.

On EL7 we need now the centos-ovirt-common-* repositories. Add both
testing and release repositories so we always consume the latest builds
in the CI.

Change-Id: I909354ba368284653ac991dc8121f2be367db4e8
Signed-off-by: Nir Soffer <nsoffer@redhat.com>