tests: Add failing tests for fsyncPath
Add failing tests synching a file and directory, and non-existing file.
IOProcessClient.fsyncPath() should return None, and raise OSError if the
path does not exists.

We don't have an easy way to test that we actually called fsync() on the
given path, so the tests focus on the return value and error handling.

Change-Id: Id854cc0356ea0b6b57e22875d08dab9141e7a01f
Bug-Url: https://bugzilla.redhat.com/1536261
Signed-off-by: Nir Soffer <nsoffer@redhat.com>