ioprocess: Wait until writefile completes
Use fsync() to ensure that the transfer completed and file metadata was
updated.

Without fsync(), some or all of the data or file metadata may be lost if
the host or storage had a fatal error before all data was transferred to
storage. This also ensures that other hosts in the cluster can see the
data when the write completes.

Change-Id: Ib0ba8b6f84a286cb0d23a665123e4b0fe1989f66
Bug-Url: https://bugzilla.redhat.com/1535429
Signed-off-by: Nir Soffer <nsoffer@redhat.com>