spec: Make python2-ioprocess arch specific
The python bindins depend on the location of /usr/libexec directory, and
on the location of the taskset command. We get both using autoconf and
generate a python module. This means that we must generate the python
module on every architecture and the package cannot be noarch package.

We will check later if we can remove the autoconf generated module and
make the package noarch again.

Change-Id: Ifabb55559113c8bbb47cacf74eb44bda9e14563f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>