readfile: Fix error handling in read loop
If the first read() failed, we were returning an empty string instead of
failing with OSError.

Change the read loop to read up to st_size bytes from the file, instead
of stopping when read() returns 0, since once we reached the end of an
unaligned file, the next read() will fail with EINVAL instead of
returning 0.

Change-Id: Ifee374e1cc04c2241ad6ad97d6fe8ad145118be3
Bug-Url: https://bugzilla.redhat.com/1726834
Signed-off-by: Nir Soffer <nsoffer@redhat.com>