ioprocess: Introduce probe_block_size
Introduce probe_block_size(), accepting a directory path, and returning
the block size that can be used for direct I/O in the directory.

Probing is done using a temporary unnamed file, so multiple hosts can
probe the same mount concurrently without triggering a crash in Gluster
fuse mount[1].

The temporary file is created using a random UUID and the O_EXCL flag,
ensuring that every call uses a new temporary file. It is also deleted
immediately, so another call is unlikely to fail trying to use the same
file.

We could simplify the implementation and avoid possible leftover probe
file using O_TMPFILE, but it seems that FUSE does not support it[2]
(open() fails with ENOTSUP).

[1] https://bugzilla.redhat.com/1751722
[2] https://sourceforge.net/p/fuse/mailman/message/33646980/

Change-Id: Ia6dac22c5415d0711c04c7710c38d86d4d43f3e8
Bug-Url: https://bugzilla.redhat.com/1753901
Signed-off-by: Nir Soffer <nsoffer@redhat.com>