ioprocess: Fix handling of large messages
When ResponseReader.process() returns False, it means that it did not
finish reading the message and it should be called again when the file
descriptor is readable. But the I/O thread exited, killing the
underlying ioprocess and starting a new one.

Fix by continuing with handling the next events. Now we process the
response again until the entire response is received.

The issue is exposed in python 3.7. For some reason this does not
happens on python 3.6.

Change-Id: I2e2da9c9c6dc4e844478265f09df874cd44530c3
Signed-off-by: Nir Soffer <nsoffer@redhat.com>