python: Do not log errors when polling fail
Due to the way that systemd terminates processes in the same cgroup
during vdsm shutdown, ioprocess is always killed before vdsm can shut it
down cleanly. When IOProcess client discover this it logs an error:

    ERROR (56ecc03c-4bb5-4792-8971-3c51ea924d2e) Communication thread failed
    Traceback (most recent call last):
       File "/usr/lib/python3.6/site-packages/ioprocess/__init__.py", line 114, in _communicate
         raise Exception("FD closed")
     Exception: FD closed

When people see this error in the logs, they assumes that ioprocess is
the reason for their issues, and they ask ioprocess maintainer to help
with their issue which is not related in any way to ioprocess.

Add special PollError, and log info message with the detail of the
error, to make sure we don't loose any detail if this was not a normal
shutdown.

Here is example log:

    INFO    (ioprocess/999116) [IOProcessClient] (ioprocess-0) Poll error 16 on fd 18

Change-Id: Ic9a7e674097e6bf5130ec6e0ea34be58d1fd36c0
Related-To: https://bugzilla.redhat.com/1851893
Signed-off-by: Nir Soffer <nsoffer@redhat.com>