python: Use module logger for client logs
The communication thread does not have a reference to the client object,
but it depends on the client for logging. Theoretically the real_ioproc
can be None when the communication thread try to log using the client
logger, logging unhelpful tracebacks.

Fix the issue by moving IOProcess._log to the module. The communication
thread keeps now the name of the client and use it to log messages for
the client to the module logger.

IOProcess._sublog is kept in the client class since it is used only from
single method, and it is useful for mocking the client logger during the
tests.

Change-Id: I82b9514281cd3780fd877a72e6fe5b1c9c5e6bc6
Signed-off-by: Nir Soffer <nsoffer@redhat.com>